+++ app/Http/Controllers/HomeController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Post;
-use App\Tag;
+use App\Models\Post;
+use App\Models\Tag;
 
 class HomeController extends Controller
 {
