+++ app/Models/Like.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
 class Like extends Model
 {
-	public $timestamps = false;
+    public $timestamps = false;
 
-	protected $fillable = [
-		'user_id',
-	];
-	
+    protected $fillable = [
+        'user_id',
+    ];
+    
     public function likeable()
     {
         return $this->morphTo();
