+++ app/Models/Tag.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use App\Models\Post;
+
+class Tag extends Model
+{
+    protected $fillable = [
+        'name'
+    ];
+
+    public function posts(){
+        return $this->belongsToMany(Post::class, 'post_tag', 'tag_id', 'post_id');
+    }
+}
