--- app/Tag.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Tag extends Model
-{
-	protected $fillable = [
-        'name'
-    ];
-
-    public function posts(){
-    	return $this->belongsToMany('App\Post', 'post_tag', 'tag_id', 'post_id');
-    }
-}
