+++ database/factories/ModelFactory.php
@@ -1,18 +1,8 @@
 <?php
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| Here you may define all of your model factories. Model factories give
-| you a convenient way to create models for testing and seeding your
-| database. Just tell the factory how a default model should look.
-|
-*/
+use App\Models\User;
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
-$factory->define(App\User::class, function (Faker\Generator $faker) {
+$factory->define(User::class, function (Faker\Generator $faker) {
     static $password;
 
     return [
@@ -21,14 +11,5 @@ $factory->define(App\User::class, function (Faker\Generator $faker) {
         'password' => $password ?: $password = bcrypt('secret'),
         'remember_token' => str_random(10),
         'role_id' => config('roles.basic')
-    ];
-});
-
-$factory->define(App\Post::class, function (Faker\Generator $faker) {
-    return [
-        'title' => $faker->sentence,
-        'content' => $faker->text,
-        'published' => config('post.review'),
-        'user_id' => factory(App\User::class)->create()->id
     ];
 });
