+++ database/seeds/CreateUserSeeder.php
@@ -2,7 +2,7 @@
 
 use Illuminate\Database\Seeder;
 
-use App\User;
+use App\Models\User;
 
 class CreateUserSeeder extends Seeder
 {
